+++ proj2/udp_stream_common.h
@@ -1,8 +1,8 @@
 #ifndef CS2520_EX2_UDP_STREAM
 #define CS2520_EX2_UDP_STREAM
-#define RESET   "\033[0m"
-#define YELLOW  "\033[33m"
-#define BOLDGREEN   "\033[1m\033[32m"
+#define RESET "\033[0m"
+#define YELLOW "\033[33m"
+#define BOLDGREEN "\033[1m\033[32m"
 #define TARGET_RATE 8 /* rate to send at, in Mbps */
 #define MAX_DATA_LEN 1300
 #define REPORT_SEC 5 /* print status every REPORT_SEC seconds */
@@ -19,23 +19,26 @@ typedef chrono::steady_clock Time;
 typedef chrono::milliseconds MS;
 typedef chrono::duration<float> Fsec;
 
-struct stream_pkt {
+struct stream_pkt
+{
     int32_t seq;
     int32_t ts_sec;
     int32_t ts_usec;
     char data[MAX_DATA_LEN];
 };
 
-struct net_pkt {
+struct net_pkt
+{
     Tp senderTS;
-    unsigned long long int seq;
+    int32_t seq;
     bool is_end;
     char data[MAX_DATA_LEN];
 } __attribute__((packed));
 
-struct ack_pkt {
-    unsigned long long int seq;
+struct ack_pkt
+{
+    int32_t seq;
     bool is_nack;
-} __attribute__ ((packed));
+} __attribute__((packed));
 
 #endif
